A bit of tweaking on Releasenotes.txt
I like to mention the most important features first (thus the profile is
added to the introductory chapter).

Moving word 'also' to different place on the GPS paragraph to indicate
that map is not the only way of inserting GPS coordinates.

JDivelog import was in 2.1 release so removed from here (it is actually
written JDiveLog - with capital L - should probably be fixed in user
manual).

Signed-off-by: Miika Turkia <miika.turkia@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>