Better editing of existing dives
This change makes it possible to edit date and time of a dive from the
normal dive edit dialog (it turns the label at the top of the dialog into
a button and allows the user to change the time from there - this seems
much more obvious than the 'right-click in the divelist' approach).

Additionally, if the dive has only one divecomputer and no samples (which
likely means that the dive was manually entered in Subsurface or imported
from another divelog where it was manually entered) then we allow editing
of duration, max depth and avg depth as well.

This fixes #75  (well, with the exception of water temperature.... we
don't allow editing of that at all... I guess that should become a new
trac item...)

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>