Fix scaling problem in scrolled edit dialog
With this change resizing the dive edit dialog once again has its content
resize as well. So now we have the best of both worlds: if the content
grows we add scroll bars (so things stay usable on small screens), yet
when we resize the window, the notes area gets nice and big (which is
great on big screens).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>