Fix bug in colour print initialisation.
The callback function for switching colour printing on/off was
registered after the initial activation of the button. This meant the
first attempt to print was compromised. We move the registration so that
printing works as the user specifies from the first attempt.

Tested by printing both B/W and colour images to file with minimal mouse
clicks.

Spotted by Salva - thanks.

Signed-off-by: Amit Chaudhuri <amit.k.chaudhuri@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>